+++ controlplane/src/core/bufservices/PlatformService.ts
@@ -1,3 +1,4 @@
+import crypto from 'node:crypto';
 import { PlainMessage } from '@bufbuild/protobuf';
 import { ServiceImpl } from '@connectrpc/connect';
 import { EnumStatusCode } from '@wundergraph/cosmo-connect/dist/common/common_pb';
@@ -5,6 +6,7 @@ import { GetConfigResponse } from '@wundergraph/cosmo-connect/dist/node/v1/node_
 import { OrganizationEventName, PlatformEventName } from '@wundergraph/cosmo-connect/dist/notifications/events_pb';
 import { PlatformService } from '@wundergraph/cosmo-connect/dist/platform/v1/platform_connect';
 import {
+  AnalyticsConfig,
   CheckFederatedGraphResponse,
   CheckSubgraphSchemaResponse,
   CompositionError,
@@ -13,11 +15,15 @@ import {
   CreateFederatedGraphTokenResponse,
   CreateFederatedSubgraphResponse,
   CreateIntegrationResponse,
+  CreateOIDCProviderResponse,
   CreateOrganizationWebhookConfigResponse,
+  DateRange,
   DeleteAPIKeyResponse,
   DeleteFederatedGraphResponse,
   DeleteFederatedSubgraphResponse,
   DeleteIntegrationResponse,
+  DeleteOIDCProviderResponse,
+  DeleteOrganizationResponse,
   DeleteRouterTokenResponse,
   FixSubgraphSchemaResponse,
   ForceCheckSuccessResponse,
@@ -27,6 +33,7 @@ import {
   GetCheckOperationsResponse,
   GetCheckSummaryResponse,
   GetChecksByFederatedGraphNameResponse,
+  GetClientsResponse,
   GetDashboardAnalyticsViewResponse,
   GetFederatedGraphByNameResponse,
   GetFederatedGraphChangelogResponse,
@@ -34,10 +41,13 @@ import {
   GetFederatedGraphsResponse,
   GetFieldUsageResponse,
   GetGraphMetricsResponse,
+  GetLatestValidSubgraphSDLByNameResponse,
   GetMetricsErrorRateResponse,
+  GetOIDCProviderResponse,
   GetOperationContentResponse,
   GetOrganizationIntegrationsResponse,
   GetOrganizationMembersResponse,
+  GetOrganizationRequestsCountResponse,
   GetOrganizationWebhookConfigsResponse,
   GetOrganizationWebhookMetaResponse,
   GetRouterTokensResponse,
@@ -46,34 +56,37 @@ import {
   GetTraceResponse,
   InviteUserResponse,
   IsGitHubAppInstalledResponse,
+  LeaveOrganizationResponse,
   MigrateFromApolloResponse,
   PublishFederatedSubgraphResponse,
+  PublishPersistedOperationsResponse,
+  PublishedOperation,
+  PublishedOperationStatus,
   RemoveInvitationResponse,
   RequestSeriesItem,
   UpdateFederatedGraphResponse,
   UpdateIntegrationConfigResponse,
+  UpdateOrgMemberRoleResponse,
   UpdateOrganizationDetailsResponse,
   UpdateOrganizationWebhookConfigResponse,
   UpdateSubgraphResponse,
   WhoAmIResponse,
-  CreateOIDCProviderResponse,
-  GetOIDCProviderResponse,
-  DeleteOIDCProviderResponse,
-  LeaveOrganizationResponse,
-  DeleteOrganizationResponse,
-  UpdateOrgMemberRoleResponse,
-  GetLatestValidSubgraphSDLByNameResponse,
 } from '@wundergraph/cosmo-connect/dist/platform/v1/platform_pb';
 import { OpenAIGraphql, isValidUrl } from '@wundergraph/cosmo-shared';
-import { parse } from 'graphql';
+import { formatISO, subHours } from 'date-fns';
+import { DocumentNode, buildASTSchema, parse } from 'graphql';
+import { validate } from 'graphql/validation/index.js';
 import { uid } from 'uid';
-import { GraphApiKeyDTO, GraphApiKeyJwtPayload } from '../../types/index.js';
+import { GraphApiKeyDTO, GraphApiKeyJwtPayload, PublishedOperationData } from '../../types/index.js';
 import { Composer } from '../composition/composer.js';
 import { buildSchema, composeSubgraphs } from '../composition/composition.js';
 import { getDiffBetweenGraphs } from '../composition/schemaCheck.js';
 import { signJwt } from '../crypto/jwt.js';
 import { FederatedGraphRepository } from '../repositories/FederatedGraphRepository.js';
 import { GitHubRepository } from '../repositories/GitHubRepository.js';
+import { GraphCompositionRepository } from '../repositories/GraphCompositionRepository.js';
+import { OidcRepository } from '../repositories/OidcRepository.js';
+import { OperationsRepository } from '../repositories/OperationsRepository.js';
 import { OrganizationRepository } from '../repositories/OrganizationRepository.js';
 import { SchemaCheckRepository } from '../repositories/SchemaCheckRepository.js';
 import { SubgraphRepository } from '../repositories/SubgraphRepository.js';
@@ -81,16 +94,18 @@ import { UserRepository } from '../repositories/UserRepository.js';
 import { AnalyticsDashboardViewRepository } from '../repositories/analytics/AnalyticsDashboardViewRepository.js';
 import { AnalyticsRequestViewRepository } from '../repositories/analytics/AnalyticsRequestViewRepository.js';
 import { MetricsRepository } from '../repositories/analytics/MetricsRepository.js';
+import { MonthlyRequestViewRepository } from '../repositories/analytics/MonthlyRequestViewRepository.js';
 import { TraceRepository } from '../repositories/analytics/TraceRepository.js';
 import { UsageRepository } from '../repositories/analytics/UsageRepository.js';
 import type { RouterOptions } from '../routes.js';
 import { ApiKeyGenerator } from '../services/ApiGenerator.js';
 import ApolloMigrator from '../services/ApolloMigrator.js';
+import OidcProvider from '../services/OidcProvider.js';
 import {
   InspectorOperationResult,
+  InspectorSchemaChange,
   SchemaUsageTrafficInspector,
   collectOperationUsageStats,
-  InspectorSchemaChange,
 } from '../services/SchemaUsageTrafficInspector.js';
 import Slack from '../services/Slack.js';
 import {
@@ -100,11 +115,9 @@ import {
   isValidLabelMatchers,
   isValidLabels,
   isValidOrganizationSlug,
+  validateDateRanges,
 } from '../util.js';
 import { FederatedGraphSchemaUpdate, OrganizationWebhookService } from '../webhooks/OrganizationWebhookService.js';
-import { OidcRepository } from '../repositories/OidcRepository.js';
-import OidcProvider from '../services/OidcProvider.js';
-import { GraphCompositionRepository } from '../repositories/GraphCompositionRepository.js';
 
 export default function (opts: RouterOptions): Partial<ServiceImpl<typeof PlatformService>> {
   return {
@@ -155,8 +168,6 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
           labelMatchers: req.labelMatchers,
           routingUrl: req.routingUrl,
         });
-
-        await fedGraphRepo.createConfig(federatedGraph.id, 7);
 
         const subgraphs = await subgraphRepo.listByFederatedGraph(req.name, {
           published: true,
@@ -654,12 +665,12 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
         }
 
         for (const composition of result.compositions) {
-          const graphConfig = await fedGraphRepo.getConfig(composition.id);
+          const orgLimits = await orgRepo.getOrganizationLimits({ organizationID: authContext.organizationId });
 
           await schemaCheckRepo.createCheckedFederatedGraph(
             schemaCheckID,
             composition.id,
-            graphConfig.trafficCheckDays,
+            orgLimits.breakingChangeRetentionLimit,
           );
 
           // We collect composition errors for all federated graphs
@@ -675,12 +686,12 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
           // We don't collect operation usage when we have composition errors or
           // when we don't have any inspectable changes. That means any breaking change is really breaking
           if (composition.errors.length === 0 && isInspectable && inspectorChanges.length > 0) {
-            if (graphConfig.trafficCheckDays <= 0) {
+            if (orgLimits.breakingChangeRetentionLimit <= 0) {
               continue;
             }
 
             const result = await trafficInspector.inspect(inspectorChanges, {
-              daysToConsider: graphConfig.trafficCheckDays,
+              daysToConsider: orgLimits.breakingChangeRetentionLimit,
               federatedGraphId: composition.id,
               organizationId: authContext.organizationId,
             });
@@ -1024,6 +1035,7 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
       return handleError<PlainMessage<GetFederatedGraphChangelogResponse>>(logger, async () => {
         const authContext = await opts.authenticator.authenticate(ctx.requestHeader);
         const fedgraphRepo = new FederatedGraphRepository(opts.db, authContext.organizationId);
+        const orgRepo = new OrganizationRepository(opts.db);
 
         const federatedGraph = await fedgraphRepo.byName(req.name);
         if (!federatedGraph) {
@@ -1047,10 +1059,17 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
           };
         }
 
+        const orgLimits = await orgRepo.getOrganizationLimits({ organizationID: authContext.organizationId });
+
+        const { dateRange } = validateDateRanges({
+          limit: orgLimits.changelogDataRetentionLimit,
+          dateRange: req.dateRange,
+        });
+
         const result = await fedgraphRepo.fetchFederatedGraphChangelog(
           federatedGraph.targetId,
           req.pagination,
-          req.dateRange,
+          dateRange!,
         );
 
         if (!result) {
@@ -1082,6 +1101,8 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
       return handleError<PlainMessage<GetChecksByFederatedGraphNameResponse>>(logger, async () => {
         const authContext = await opts.authenticator.authenticate(ctx.requestHeader);
         const fedgraphRepo = new FederatedGraphRepository(opts.db, authContext.organizationId);
+        const subgraphRepo = new SubgraphRepository(opts.db, authContext.organizationId);
+        const orgRepo = new OrganizationRepository(opts.db);
 
         const federatedGraph = await fedgraphRepo.byName(req.name);
         if (!federatedGraph) {
@@ -1095,13 +1116,22 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
           };
         }
 
-        const subgraphRepo = new SubgraphRepository(opts.db, authContext.organizationId);
+        const orgLimits = await orgRepo.getOrganizationLimits({ organizationID: authContext.organizationId });
+
+        const { dateRange } = validateDateRanges({
+          limit: orgLimits.breakingChangeRetentionLimit,
+          dateRange: {
+            start: req.startDate,
+            end: req.endDate,
+          } as DateRange,
+        });
+
         const checksData = await subgraphRepo.checks({
           federatedGraphName: req.name,
           limit: req.limit,
           offset: req.offset,
-          startDate: req.startDate,
-          endDate: req.endDate,
+          startDate: dateRange!.start,
+          endDate: dateRange!.end,
         });
         const totalChecksCount = await subgraphRepo.getChecksCount({ federatedGraphName: req.name });
 
@@ -1391,6 +1421,9 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
             },
           };
         }
+
+        const blobStorageDirectory = `${authContext.organizationId}/${federatedGraph.id}`;
+        await opts.blobStorage.removeDirectory(blobStorageDirectory);
 
         const subgraphsTargetIDs: string[] = [];
         const subgraphs = await subgraphRepo.listByFederatedGraph(req.name);
@@ -1720,6 +1753,7 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
         const authContext = await opts.authenticator.authenticate(ctx.requestHeader);
         const analyticsRepo = new AnalyticsRequestViewRepository(opts.chClient);
         const fedGraphRepo = new FederatedGraphRepository(opts.db, authContext.organizationId);
+        const orgRepo = new OrganizationRepository(opts.db);
 
         const graph = await fedGraphRepo.byName(req.federatedGraphName);
         if (!graph) {
@@ -1731,11 +1765,22 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
           };
         }
 
+        const orgLimits = await orgRepo.getOrganizationLimits({ organizationID: authContext.organizationId });
+        const { range, dateRange } = validateDateRanges({
+          limit: orgLimits.tracingRetentionLimit,
+          range: req.config?.range,
+          dateRange: req.config?.dateRange,
+        });
+
         return {
           response: {
             code: EnumStatusCode.OK,
           },
-          view: await analyticsRepo.getView(authContext.organizationId, graph.id, req.name, req.config),
+          view: await analyticsRepo.getView(authContext.organizationId, graph.id, req.name, {
+            ...req.config,
+            range,
+            dateRange,
+          } as AnalyticsConfig),
         };
       });
     },
@@ -1805,6 +1850,7 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
         const authContext = await opts.authenticator.authenticate(ctx.requestHeader);
         const repo = new MetricsRepository(opts.chClient);
         const fedGraphRepo = new FederatedGraphRepository(opts.db, authContext.organizationId);
+        const orgRepo = new OrganizationRepository(opts.db);
 
         const graph = await fedGraphRepo.byName(req.federatedGraphName);
         if (!graph) {
@@ -1817,9 +1863,16 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
           };
         }
 
-        const view = await repo.getMetricsView({
+        const orgLimits = await orgRepo.getOrganizationLimits({ organizationID: authContext.organizationId });
+        const { range, dateRange } = validateDateRanges({
+          limit: orgLimits.analyticsRetentionLimit,
           range: req.range,
           dateRange: req.dateRange,
+        });
+
+        const view = await repo.getMetricsView({
+          range,
+          dateRange,
           filters: req.filters,
           organizationId: authContext.organizationId,
           graphId: graph.id,
@@ -1852,6 +1905,7 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
         const authContext = await opts.authenticator.authenticate(ctx.requestHeader);
         const repo = new MetricsRepository(opts.chClient);
         const fedGraphRepo = new FederatedGraphRepository(opts.db, authContext.organizationId);
+        const orgRepo = new OrganizationRepository(opts.db);
 
         const graph = await fedGraphRepo.byName(req.federatedGraphName);
         if (!graph) {
@@ -1864,9 +1918,16 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
           };
         }
 
-        const metrics = await repo.getErrorsView({
+        const orgLimits = await orgRepo.getOrganizationLimits({ organizationID: authContext.organizationId });
+        const { range, dateRange } = validateDateRanges({
+          limit: orgLimits.analyticsRetentionLimit,
           range: req.range,
           dateRange: req.dateRange,
+        });
+
+        const metrics = await repo.getErrorsView({
+          range,
+          dateRange,
           filters: req.filters,
           organizationId: authContext.organizationId,
           graphId: graph.id,
@@ -3883,6 +3944,198 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
           response: {
             code: EnumStatusCode.OK,
           },
+        };
+      });
+    },
+
+    publishPersistedOperations: (req, ctx) => {
+      /**
+       * Receives a federated graph name and a list of persisted operation contents.
+       * First, it validates that the graph exists and all the operations are valid,
+       * then it stores them. Additionally, if the provided client name for registering
+       * the operations has never been seen before, we create an entry in the database
+       * with it.
+       */
+      const logger = opts.logger.child({
+        service: ctx.service.typeName,
+        method: ctx.method.name,
+      });
+
+      return handleError<PlainMessage<PublishPersistedOperationsResponse>>(logger, async () => {
+        const authContext = await opts.authenticator.authenticate(ctx.requestHeader);
+        if (!authContext.hasWriteAccess) {
+          return {
+            response: {
+              code: EnumStatusCode.ERR,
+              details: `The user doesnt have the permissions to perform this operation`,
+            },
+            operations: [],
+          };
+        }
+        const userId = authContext.userId;
+        if (!userId) {
+          return {
+            response: {
+              code: EnumStatusCode.ERROR_NOT_AUTHENTICATED,
+              details: `User not found in the authentication context`,
+            },
+            operations: [],
+          };
+        }
+        const organizationId = authContext.organizationId;
+        const federatedGraphRepo = new FederatedGraphRepository(opts.db, organizationId);
+
+        // Validate everything before we update any data
+        const schema = await federatedGraphRepo.getLatestValidSchemaVersion(req.fedGraphName);
+        const federatedGraph = await federatedGraphRepo.byName(req.fedGraphName);
+        if (!schema?.schema || federatedGraph === undefined) {
+          return {
+            response: {
+              code: EnumStatusCode.ERR_NOT_FOUND,
+              details: `Federated graph '${req.fedGraphName}' does not exist`,
+            },
+            operations: [],
+          };
+        }
+        const graphAST = parse(schema.schema);
+        const graphSchema = buildASTSchema(graphAST);
+        for (const operationContents of req.operations) {
+          let opAST: DocumentNode;
+          try {
+            opAST = parse(operationContents);
+          } catch (e: any) {
+            return {
+              response: {
+                code: EnumStatusCode.ERR,
+                details: `Operation ${operationContents} is not valid: ${e}`,
+              },
+              operations: [],
+            };
+          }
+          const errors = validate(graphSchema, opAST, undefined, { maxErrors: 1 });
+          if (errors.length > 0) {
+            const errorDetails = errors.map((e) => `${e.toString()}`).join(', ');
+            return {
+              response: {
+                code: EnumStatusCode.ERR,
+                details: `Operation "${operationContents}" is not valid: ${errorDetails}`,
+              },
+              operations: [],
+            };
+          }
+        }
+        const operationsRepo = new OperationsRepository(opts.db, federatedGraph.id);
+        let clientId: string;
+        try {
+          clientId = await operationsRepo.registerClient(req.clientName, userId);
+        } catch (e: any) {
+          const message = e instanceof Error ? e.message : e.toString();
+          return {
+            response: {
+              code: EnumStatusCode.ERR,
+              details: `Could not register client "${req.clientName}": ${message}`,
+            },
+            operations: [],
+          };
+        }
+        const operations: PublishedOperation[] = [];
+        const updatedOperations = [];
+        // Retrieve the operations that have already been published
+        const operationsResult = await operationsRepo.getPersistedOperations();
+        const operationHashes = new Set(operationsResult.map((op) => op.hash));
+        for (const operationContents of req.operations) {
+          const operationHash = crypto.createHash('sha256').update(operationContents).digest('hex');
+          const path = `${organizationId}/${federatedGraph.id}/operations/${req.clientName}/${operationHash}.json`;
+          updatedOperations.push({
+            hash: operationHash,
+            filePath: path,
+          });
+          let status: PublishedOperationStatus;
+          if (operationHashes.has(operationHash)) {
+            status = PublishedOperationStatus.UP_TO_DATE;
+          } else {
+            const data: PublishedOperationData = {
+              version: 1,
+              body: operationContents,
+            };
+            opts.blobStorage.putObject(path, Buffer.from(JSON.stringify(data), 'utf8'));
+            status = PublishedOperationStatus.CREATED;
+          }
+          operations.push(
+            new PublishedOperation({
+              hash: operationHash,
+              status,
+            }),
+          );
+        }
+
+        await operationsRepo.updatePersistedOperations(clientId, userId, updatedOperations);
+
+        return {
+          response: {
+            code: EnumStatusCode.OK,
+          },
+          operations,
+        };
+      });
+    },
+
+    getClients: (req, ctx) => {
+      const logger = opts.logger.child({
+        service: ctx.service.typeName,
+        method: ctx.method.name,
+      });
+
+      return handleError<PlainMessage<GetClientsResponse>>(logger, async () => {
+        const authContext = await opts.authenticator.authenticate(ctx.requestHeader);
+        const fedRepo = new FederatedGraphRepository(opts.db, authContext.organizationId);
+        const federatedGraph = await fedRepo.byName(req.fedGraphName);
+        if (!federatedGraph) {
+          return {
+            response: {
+              code: EnumStatusCode.ERR_NOT_FOUND,
+              details: `Federated graph '${req.fedGraphName}' does not exist`,
+            },
+            clients: [],
+          };
+        }
+        const operationsRepo = new OperationsRepository(opts.db, federatedGraph.id);
+        const clients = await operationsRepo.getRegisteredClients();
+
+        return {
+          response: {
+            code: EnumStatusCode.OK,
+          },
+          clients,
+        };
+      });
+    },
+
+    getOrganizationRequestsCount: (req, ctx) => {
+      const logger = opts.logger.child({
+        service: ctx.service.typeName,
+        method: ctx.method.name,
+      });
+
+      return handleError<PlainMessage<GetOrganizationRequestsCountResponse>>(logger, async () => {
+        const authContext = await opts.authenticator.authenticate(ctx.requestHeader);
+
+        if (!opts.chClient) {
+          return {
+            response: {
+              code: EnumStatusCode.ERR_ANALYTICS_DISABLED,
+            },
+            count: BigInt(0),
+          };
+        }
+        const monthlyRequestsRepo = new MonthlyRequestViewRepository(opts.chClient);
+        const count = await monthlyRequestsRepo.getMonthlyRequestCount(authContext.organizationId);
+
+        return {
+          response: {
+            code: EnumStatusCode.OK,
+          },
+          count: BigInt(count),
         };
       });
     },