+++ controlplane/src/core/util.ts
@@ -3,6 +3,8 @@ import pino from 'pino';
 import { joinLabel, splitLabel } from '@wundergraph/cosmo-shared';
 import { uid } from 'uid/secure';
 import { GraphQLSubscriptionProtocol } from '@wundergraph/cosmo-connect/dist/common/common_pb';
+import { DateRange } from '@wundergraph/cosmo-connect/dist/platform/v1/platform_pb';
+import { formatISO, subHours } from 'date-fns';
 import { Label, ResponseMessage } from '../types/index.js';
 import { MemberRole } from '../db/models.js';
 import { isAuthenticationError, isPublicError } from './errors/errors.js';
@@ -187,4 +189,37 @@ export const isValidOrganizationSlug = (slug: string): boolean => {
   }
 
   return true;
+};
+
+export const validateDateRanges = ({
+  limit,
+  range,
+  dateRange,
+}: {
+  limit: number;
+  range?: number;
+  dateRange?: DateRange;
+}): { range: number | undefined; dateRange: DateRange | undefined } => {
+  let validatedRange: number | undefined = range;
+  const validatedDateRange: DateRange | undefined = dateRange;
+
+  if (validatedRange && validatedRange > limit * 24) {
+    validatedRange = limit * 24;
+  }
+
+  if (validatedDateRange) {
+    const startDate = new Date(validatedDateRange.start);
+    const endDate = new Date(validatedDateRange.end);
+    if (startDate < subHours(new Date(), limit * 24)) {
+      validatedDateRange.start = formatISO(subHours(new Date(), limit * 24));
+    }
+    if (endDate > new Date()) {
+      validatedDateRange.end = formatISO(new Date());
+    }
+  }
+
+  return {
+    range: validatedRange,
+    dateRange: validatedDateRange,
+  };
 };