+++ controlplane/src/types/index.ts
@@ -267,3 +267,41 @@ export interface SlackAccessTokenResponse {
   slackChannelName: string;
   webhookURL: string;
 }
+
+export interface OrganizationLimitsDTO {
+  analyticsRetentionLimit: number;
+  tracingRetentionLimit: number;
+  changelogDataRetentionLimit: number;
+  breakingChangeRetentionLimit: number;
+  traceSamplingRateLimit: number;
+  requestsLimit: number;
+}
+
+export interface ClientDTO {
+  id: string;
+  name: string;
+  createdAt: string;
+  createdBy: string;
+  lastUpdatedAt: string;
+  lastUpdatedBy: string;
+}
+
+export interface PersistedOperationDTO {
+  id: string;
+  hash: string;
+  filePath: string;
+  createdAt: string;
+  createdBy: string;
+  lastUpdatedAt: string;
+  lastUpdatedBy: string;
+}
+
+export interface PublishedOperationData {
+  version: 1;
+  body: string;
+}
+
+export interface UpdatedPersistedOperation {
+  hash: string;
+  filePath: string;
+}