+++ controlplane/test/federated-graph.test.ts
@@ -19,7 +19,7 @@ import {
 } from '../src/core/test-util';
 import Keycloak from '../src/core/services/Keycloak';
 import { MockPlatformWebhookService } from '../src/core/webhooks/PlatformWebhookService';
-import { SetupTest } from './test-util';
+import { InMemoryBlobStorage, SetupTest } from './test-util';
 
 let dbname = '';
 
@@ -122,6 +122,7 @@ describe('Federated Graph', (ctx) => {
           clientSecret: '',
         },
         keycloakApiUrl: apiUrl,
+        blobStorage: new InMemoryBlobStorage(),
       }),
     });
 
@@ -224,6 +225,7 @@ describe('Federated Graph', (ctx) => {
           clientSecret: '',
         },
         keycloakApiUrl: apiUrl,
+        blobStorage: new InMemoryBlobStorage(),
       }),
     });
 
@@ -326,6 +328,7 @@ describe('Federated Graph', (ctx) => {
           clientSecret: '',
         },
         keycloakApiUrl: apiUrl,
+        blobStorage: new InMemoryBlobStorage(),
       }),
     });
 
@@ -457,6 +460,7 @@ describe('Federated Graph', (ctx) => {
           clientSecret: '',
         },
         keycloakApiUrl: apiUrl,
+        blobStorage: new InMemoryBlobStorage(),
       }),
     });